# Optimize predictors 0-9 in `lossless_transform`