- Rewrite predictor transform 1 to be auto-vectorized.
- Add an average function which performs average with bitwise operations.
  - LLVM is capable of doing this optimization and does it for the `average2` calls which are left unchanged, but it's not able to automatically perform it for every transform we have.
- Add asserts to remove some saturating arithmetic instructions.

Predictors 10-13 were much slower when rewritten following this pattern.

Related
#95, #96

---

Rust 1.82 (2024-10-17) upgraded to LLVM 19 which auto-vectorizes the old `while` constructions. This allowed for rewriting predictor 1 to be 3x faster.

1, 5, 6, and 7 are the decisive improvements in the benchmarks.

before
```

test lossless_transform::benches::predictor00     ... bench:          98.81 ns/iter (+/- 2.49) = 10408 MB/s
test lossless_transform::benches::predictor01     ... bench:         308.75 ns/iter (+/- 19.58) = 3311 MB/s
test lossless_transform::benches::predictor02     ... bench:          42.42 ns/iter (+/- 1.10) = 24285 MB/s
test lossless_transform::benches::predictor03     ... bench:          41.83 ns/iter (+/- 1.23) = 24878 MB/s
test lossless_transform::benches::predictor04     ... bench:          39.81 ns/iter (+/- 3.22) = 26153 MB/s
test lossless_transform::benches::predictor05     ... bench:         711.89 ns/iter (+/- 17.23) = 1434 MB/s
test lossless_transform::benches::predictor06     ... bench:       1,519.71 ns/iter (+/- 50.47) = 671 MB/s
test lossless_transform::benches::predictor07     ... bench:         518.31 ns/iter (+/- 4.46) = 1969 MB/s
test lossless_transform::benches::predictor08     ... bench:          70.32 ns/iter (+/- 8.86) = 14571 MB/s
test lossless_transform::benches::predictor09     ... bench:          65.92 ns/iter (+/- 1.98) = 15692 MB/s

```

after
```

test lossless_transform::benches::predictor00     ... bench:          91.49 ns/iter (+/- 1.01) = 11208 MB/s
test lossless_transform::benches::predictor01     ... bench:          90.69 ns/iter (+/- 2.70) = 11333 MB/s
test lossless_transform::benches::predictor02     ... bench:          36.57 ns/iter (+/- 1.10) = 28333 MB/s
test lossless_transform::benches::predictor03     ... bench:          34.14 ns/iter (+/- 5.56) = 30000 MB/s
test lossless_transform::benches::predictor04     ... bench:          35.98 ns/iter (+/- 2.84) = 29142 MB/s
test lossless_transform::benches::predictor05     ... bench:         466.55 ns/iter (+/- 29.46) = 2188 MB/s
test lossless_transform::benches::predictor06     ... bench:         717.89 ns/iter (+/- 48.46) = 1422 MB/s
test lossless_transform::benches::predictor07     ... bench:         268.98 ns/iter (+/- 9.90) = 3805 MB/s
test lossless_transform::benches::predictor08     ... bench:          63.66 ns/iter (+/- 1.51) = 16190 MB/s
test lossless_transform::benches::predictor09     ... bench:          61.02 ns/iter (+/- 7.13) = 16721 MB/s

```