@fintelia 2025-07-11
I suspect this PR is working around around this regression: https://github.com/rust-lang/rust/issues/142519