@fintelia 2025-07-12
I see a ~1% end-to-end speedup from this change. At this point, roughly 20% of total decode time (of lossless images) is spent in the transforms. The rough breakdown is:
```
predictor11: 8.8% 
predictor12: 4.1%
predictor13: 1.4%
predictor1:  1.1%
predictor7:  0.8%
all others:  2.6%
```